+++ packages/ui-kit/src/components/loading-overlay/loading-overlay.tsx
@@ -1,0 +1,10 @@
+import { Spinner } from '../spinner'
+
+import s from './loading-overlay.module.scss'
+export const LoadingOverlay = () => {
+  return (
+    <div className={s.backdrop}>
+      <Spinner />
+    </div>
+  )
+}